backend/requirements.txt
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudinary==1.44.0
colorama==0.4.6
contourpy==1.3.2
cors==1.0.1
cycler==0.12.1
dj-database-url==3.0.1
Django==5.2.4
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==23.5.26
fonttools==4.58.4
future==1.0.0
gast==0.5.4
gevent==25.5.1
google-api-core==2.25.1
google-api-python-client==2.175.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.60.0
gspread==6.2.1
gunicorn==23.0.0
h11==0.16.0
httplib2==0.22.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.6
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.8
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.10.3
numpy==1.26.3
oauth2client==4.1.3
oauthlib==3.2.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.3.0
pillow==11.2.1
proto-plus==1.26.1
protobuf==4.23.4
psycopg2-binary==2.9.10
pyasn1==0.5.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.3
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
pytz==2023.3.post1
requests==2.32.4
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9.1
seaborn==0.13.2
selenium==4.33.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sqlparse==0.5.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tldextract==5.3.0
trio==0.30.0
trio-websocket==0.12.2
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.2.0
urllib3==2.1.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.14.1
wsproto==1.2.0
zope.event==5.1
zope.interface==7.2
